+++ src/pages/_app.tsx
@@ -1,8 +1,8 @@
 import type { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
 
-import { ContentWrapper } from "../components/ContentWrapper";
-import { SiteHeader } from "../components/SiteHeader";
+import { ContentWrapper } from "../components/content-wrapper";
+import { SiteHeader } from "../components/site-header";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
